auth_service/requirements.txt
Flask
redis
psycopg2-binary 